Add a possibility to change the depth of the breadcrumbs
By default, breadcrumbs show all levels up to the root page. This
might be undesirable if subtrees are considered more like
independent wikis. This commit adds the possibility to choose the
root level which is the highest level up to which the breadcrumbs
are shown. In order to use this feature, one must add a context
processor wiki.context_processors.breadcrumbs and define
WIKI_BREADCRUMBS_ROOT_LEVEL (default is 0) in settings. If these
are not set, wiki works as before.